build(lint): add curated golangci-lint (misspell off) + fix all 114 findings (backlog E1)
Wire a curated golangci-lint v2 gate and drive it to zero.

Config (.golangci.yml): enable errcheck, govet, ineffassign, staticcheck,
unconvert, unused. misspell stays OFF — it flagged French comments as
misspelled English (1097 of the original 1370 findings = noise). Tests skip
errcheck (a failed setup surfaces at the next assertion) and SA1012
(nil-context is a legitimate test pattern). cmd/iterion-desktop is excluded:
it is a Wails/cgo app behind //go:build desktop, so the core lint sees only
its untagged files in isolation and reports false "unused" on symbols the
tagged files consume.

Findings fixed (114 -> 0):
- staticcheck QF* simplifications: WriteString(Sprintf) -> Fprintf, tagged
  switches, embedded-field selectors, De Morgan, ReplaceAll.
- errcheck (production): explicit err check in the stream-json parser;
  explicit _ = on best-effort websocket keepalive deadlines (hub.go) and
  file-lock/cleanup defers; capture + log filepath.WalkDir / fsWatcher.Add
  failures in the file watcher.
- unused: drop the dead gitDiffBlobName const; //nolint the documented
  allowlist justification slot.
- ineffassign/SA4006: drop dead maxReplayed writes (the replay happens inside
  eventTerminalCatchUp); collapse the identical newStatus branches; declare
  probe transient without the ineffectual init.
- ST1005/ST1008: lowercase bash error strings, reword manifest errors to not
  end in a colon, reorder runWithSeparateStreams to (stdout, stderr, err).
- SA1019: //nolint tar.TypeRegA (needed to read legacy tar archives).

Enforcement: task lint now runs golangci-lint (devbox provides 2.12.2); a new
parallel CI `golangci` job installs the pinned binary and runs it. Add it to
the branch-protection required checks to gate merges.

task check + task test:race green; golangci-lint run: 0 issues.

+++ pkg/runview/snapshot_test.go
@@ -17,7 +17,7 @@ import (
 func evt(seq int64, t store.EventType, branch, node string, data map[string]any) *store.Event {
 	return &store.Event{
 		Seq:       seq,
-		Timestamp: time.Unix(int64(seq), 0).UTC(),
+		Timestamp: time.Unix(seq, 0).UTC(),
 		Type:      t,
 		BranchID:  branch,
 		NodeID:    node,
